sensor_calibration: override for fmu-v6c
This is a workaround for Pixhawk 6C revision 0 where the internal mag
and baro are using the external I2C bus.

In order for the mag calibration to treat the mag as an internal one, as
well as sensors_status_mag and sensors_status_baro to be correct, we
need to insert this override to change these two sensors specifically to
internal.